# openapi.yaml
openapi: 3.0.3
info:
  title: API Sổ thu chi cá nhân
  version: 2.0.0
  description: REST API cho TASK_00030. Mọi response JSON sử dụng envelope success/message/data hoặc success/error.
servers:
  - url: http://localhost:3000/api/v1
...
paths:
  /auth/register:
    post:
      tags: [Auth]
      summary: Đăng ký tài khoản
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RegisterRequest' }
      responses:
        '201': { $ref: '#/components/responses/Success' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /auth/login:
    post:
      tags: [Auth]
      summary: Đăng nhập bằng username, email hoặc số điện thoại
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [identifier, password]
              properties:
                identifier: { type: string, example: demo }
                password: { type: string, format: password, example: Demo@123 }
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Xoay vòng access/refresh token
      security: []
      requestBody: { $ref: '#/components/requestBodies/RefreshToken' }
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /auth/logout:
    post:
      tags: [Auth]
      summary: Đăng xuất và thu hồi refresh token
      requestBody: { $ref: '#/components/requestBodies/RefreshToken' }
      responses:
        '200': { $ref: '#/components/responses/Success' }
  /auth/forgot-password:
    post:
      tags: [Auth]
      summary: Gửi email đặt lại mật khẩu
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [identifier]
              properties: { identifier: { type: string } }
      responses: { '200': { $ref: '#/components/responses/Success' } }
  /auth/reset-password:
    post:
      tags: [Auth]
      summary: Đặt lại mật khẩu bằng token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, newPassword]
              properties:
                token: { type: string }
                newPassword: { type: string, format: password }
      responses: { '200': { $ref: '#/components/responses/Success' } }
  /auth/change-password:
    post:
      tags: [Auth]
      summary: Đổi mật khẩu
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [currentPassword, newPassword]
              properties:
                currentPassword: { type: string, format: password }
                newPassword: { type: string, format: password }
      responses: { '200': { $ref: '#/components/responses/Success' } }
...
  /auth/sessions/{familyId}:
    delete:
      tags: [Auth]
      summary: Thu hồi một chuỗi phiên trên thiết bị
      parameters: [{ in: path, name: familyId, required: true, schema: { type: string, format: uuid } }]
      responses: { '200': { $ref: '#/components/responses/Success' } }
...
  /auth/oauth/{provider}/start:
    get:
      tags: [Auth]
      summary: Bắt đầu đăng nhập Google hoặc GitHub
      security: []
      parameters: [{ in: path, name: provider, required: true, schema: { type: string, enum: [google, github] } }]
      responses: { '302': { description: Chuyển tới nền tảng OAuth } }
...
  /auth/verification/email/confirm:
    post:
      tags: [Auth]
      summary: Xác nhận token email
      security: []
      requestBody: { required: true, content: { application/json: { schema: { type: object, required: [token], properties: { token: { type: string } } } } } }
      responses: { '200': { $ref: '#/components/responses/Success' } }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
...
  requestBodies:
    RefreshToken:
      required: true
      content:
        application/json:
          schema:
            type: object
            required: [refreshToken]
            properties: { refreshToken: { type: string } }
  responses:
    Success:
      description: Thành công
      content: { application/json: { schema: { $ref: '#/components/schemas/SuccessEnvelope' } } }
    ValidationError:
      description: Dữ liệu không hợp lệ
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
    Unauthorized:
      description: Chưa xác thực hoặc token không hợp lệ
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
  schemas:
    SuccessEnvelope:
      type: object
      required: [success, message, data]
      properties:
        success: { type: boolean, example: true }
        message: { type: string, example: Thành công. }
        data: { nullable: true }
        meta: { type: object, nullable: true }
    ErrorEnvelope:
      type: object
      required: [success, error]
      properties:
        success: { type: boolean, example: false }
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string, example: VALIDATION_ERROR }
            message: { type: string }
            details: { nullable: true }
    RegisterRequest:
      type: object
      required: [username, password]
      properties:
        username: { type: string, minLength: 3, example: demo }
        email: { type: string, format: email, example: demo@example.com }
        phone: { type: string, example: '+84901234567' }
        password: { type: string, format: password, minLength: 10, example: MatKhauDai@2026 }
        fullName: { type: string, example: Người dùng Demo }
...
security:
  - bearerAuth: []